Migrates functions to presenters
