+++ .eslintrc.js
@@ -37,6 +37,7 @@ module.exports = {
     '@typescript-eslint/no-shadow': ['error'],
     'no-unused-vars': 'off',
     '@typescript-eslint/no-unused-vars': ['warn'],
+    'import/prefer-default-export': ['off'],
   },
   overrides: [
     {
